Code reorg to separate UI code (#16)
code reorg to separate some of the UI code
