+++ README.md
@@ -103,6 +103,7 @@ This project is licensed under the MIT License - see the [LICENSE](LICENSE) file
 ## ➕ Enhancements from base repository ➕
 - Do not need Open AI key, paid Open AI account to use the complete functionality
 - Allow users to add contextual information to provide customized responses to conversation
+- Allows usage of different models for transcription using command line arguments
 - Allow to pause audio transcription
 - Transcribe audio of any video
 - Preserve all conversation text in UI
